exp video/requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
openai-whisper>=20230314
TTS>=0.11.1
opencv-python>=4.5.0
numpy>=1.21.0
psutil>=5.8.0
ffmpeg-python>=0.2.0